add new test code
